+++ packages/app/src/kits/Buttons/index.scss
@@ -119,37 +119,6 @@
 
 /* Button Styles */
 
-.btn-secondary {
-  @include btn-core;
-  @include btn-layout;
-  @include btn-spacing;
-  @include btn-icons;
-  @include btn-disabled;
-
-  background: var(--button-secondary-background);
-  border: 1px solid var(--button-secondary-background);
-  border-radius: var(--button-border-radius-large);
-  color: var(--text-color-primary);
-
-  &.sm {
-    @include btn-small;
-
-    border-radius: var(--button-border-radius-large);
-  }
-
-  &.lg {
-    @include btn-large;
-  }
-
-  &:hover {
-    background: var(--button-hover-background);
-
-    &:disabled {
-      background: var(--button-secondary-background);
-    }
-  }
-}
-
 .btn-submit {
   @include btn-core;
   @include btn-layout;
