+++ packages/app/src/kits/Structure/PageTitle/index.tsx
@@ -4,7 +4,7 @@
 import { useEffect, useRef, useState } from 'react';
 import { faBars } from '@fortawesome/free-solid-svg-icons';
 import { PageTitleTabs } from '../PageTitleTabs';
-import { ButtonSecondary } from 'kits/Buttons/ButtonSecondary';
+import { ButtonSecondary } from 'ui-buttons';
 import { appendOrEmpty } from '@w3ux/utils';
 import type { PageTitleProps } from './types';
 import { PageTitleWrapper, ScrollableWrapper } from './Wrappers';
