+++ packages/app/src/modals/Connect/ReadOnly.tsx
@@ -22,8 +22,7 @@ import {
   ManualAccountsWrapper,
 } from './Wrappers';
 import type { ListWithInputProps } from './types';
-import { ButtonHelp, ButtonMonoInvert } from 'ui-buttons';
-import { ButtonSecondary } from 'kits/Buttons/ButtonSecondary';
+import { ButtonHelp, ButtonMonoInvert, ButtonSecondary } from 'ui-buttons';
 
 export const ReadOnly = ({ setInputOpen, inputOpen }: ListWithInputProps) => {
   const { t } = useTranslation('modals');
