+++ packages/ui-buttons/src/ButtonSecondary/index.module.scss
@@ -1,0 +1,9 @@
+// Copyright 2024 @polkadot-cloud/polkadot-staking-dashboard authors & contributors
+// SPDX-License-Identifier: GPL-3.0-only
+
+.btnSecondary {
+  background: var(--button-secondary-background);
+  border: 1px solid var(--button-secondary-background);
+  border-radius: var(--button-border-radius-large);
+  color: var(--text-color-primary);
+}
